refactor(types): reorganise instantsearch type files
