+++ src/types/index.ts
@@ -1,7 +1,16 @@
+// internal
+export * from './utils';
+
+// Algolia-related
+export * from './algoliasearch';
+export * from './results';
+
+// instantsearch-related
 export * from './instantsearch';
+export * from './middleware';
+export * from './router';
+export * from './insights';
+
+// widget-related
 export * from './connector';
 export * from './widget';
-export * from './insights';
-export * from './algoliasearch';
-export * from './middleware';
-export * from './utils';
